// src/types/schemas/playbackReplicant.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PlaybackReplicant {
	/**
	 * 0: Stopped, 1: Playing, 2: Paused
	 */
	state: 0 | 1 | 2;
	title: string;
	artist?: string;
	album?: string;
}